feat: more clear structure
