+++ src/internal/interpreter/map-interpreter.test.ts
@@ -480,7 +480,6 @@ describe('MapInterpreter', () => {
         }
       }`);
     const result = await interpreter.perform(ast);
-    console.log('res', result);
     expect(result.isOk() && result.value).toEqual(12);
   });
